refactor: restructure backend for production
